pyproject.toml
[project]
name = "tomboonedotcom"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
